Project/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid==2.12.4
Click==8.1.3
ecdsa==0.18.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.0
future==0.18.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
pycryptodome==3.15.0
pylint==2.15.0
python-jose==3.3.0
six==1.16.0
typed-ast==1.5.4
Werkzeug==2.2.2
wrapt==1.11.1
Flask-Cors==3.0.10